microservices/analytics-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pymongo==4.5.0
httpx==0.25.2
ua-parser==0.18.0
user-agents==2.2.0
vobject==0.9.6.1
python-multipart==0.0.6
pydantic==2.4.2
python-jose[cryptography]
redis==5.0.1
PyJWT